clean some errors
