+++ soen390_web/src/pages/UserProfile.jsx
@@ -5,28 +5,23 @@ import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import profilepicture from "../assets/default_picture.jpg";
 import background from "../assets/profile_background.svg";
-import AmazonLogo from "../assets/UserProfileImages/amazon-logo-square.jpg";
-import Concordia from "../assets/UserProfileImages/Concordia.png";
+import AddDocumentsDialog from "../components/AddDocumentsDialog";
 import AddEducationDialog from "../components/AddEducationDialog";
 import AddExperienceDialog from "../components/AddExperienceDialog";
 import AddSkillDialog from "../components/AddSkillDialog";
-import Footer from "../components/Footer";
-import SubFooter from "../components/SubFooter";
+import ExperienceList from "../components/ExperienceList";
 import "../styles/components/UserProfile.css";
 import ApplicationHistory from "./ApplicationHistory";
-import AddDocumentsDialog from "../components/AddDocumentsDialog";
-import ExperienceItem from "../components/ExperienceItem";
-import ExperienceList from "../components/ExperienceList";
 
-import { GetFile } from "../api/UserStorageApi";
+import { useTranslation } from "react-i18next";
 import {
+  addSkill,
   getExperience,
   getSkills,
-  addSkill,
   removeSkill,
 } from "../api/UserProfileApi";
+import { GetFile } from "../api/UserStorageApi";
 import ProfileFileItem from "../components/ProfileFileItem";
-import { useTranslation } from "react-i18next";
 
 function UserProfile(props) {
   const [enable, setEnable] = React.useState(false);
